ENG-0000 - Support Alternative Magma Locations
Also a little code reorganization and removal of deadwood
