+++ src/common/navigation/index.ts
@@ -1,5 +1,6 @@
 export * from "./al-location.dictionary";
-export * from "./al-locator.service";
 export * from "./al-locator.types";
+export * from "./al-locator.matrix";
+export * from "./al-locator.service";
 export * from "./al-route.types";
 export * from './al-navigation.types';
